+++ source/frontend/components/pages/Overview.vue
@@ -7,6 +7,7 @@
     </f7-block>
     <f7-list>
       <f7-list-item title="Hello World" link="/helloWorld/"></f7-list-item>
+      <f7-list-item title="Local Storage" link="/localStorage/"></f7-list-item>
       <f7-list-item title="Collection" link="/collection/"></f7-list-item>
       <f7-list-item title="MySQL API" link="/mysqlapi/"></f7-list-item>
       <f7-list-item title="MySQL Collection" link="/mysqlCollection/"></f7-list-item>