+++ source/frontend/composables/index.js
@@ -1,11 +1,15 @@
+import { useLocalStorage } from './LocalStorage'
 import { useCollection } from './Collection'
 import { useHelloWorld } from './HelloWorld'
 import { useMySQLAPI } from './MySQLAPI'
 import { useMySQLCollection } from './MySQLCollection'
+import { useType } from './Type'
 
 export {
+  useLocalStorage,
   useCollection,
   useHelloWorld,
   useMySQLAPI,
-  useMySQLCollection
+  useMySQLCollection,
+  useType
 }