+++ source/frontend/routes.js
@@ -1,4 +1,5 @@
 import OverviewPage from './components/pages/Overview.vue'
+import LocalStoragePage from './components/pages/LocalStorage.vue'
 import CollectionPage from './components/pages/Collection.vue'
 import HelloWorldPage from './components/pages/HelloWorld.vue'
 import MySQLAPIPage from './components/pages/MySQLAPI.vue'
@@ -8,6 +9,10 @@ export default [
   {
     path: '/',
     component: OverviewPage
+  },
+  {
+    path: '/localStorage/',
+    component: LocalStoragePage
   },
   {
     path: '/collection/',