+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [NEXT]
+
+### Added
+
+- Type to get the type of the value
+- LocalStorage to have a reactive and persistent value
+
 ## [v1.36.2]
 
 ### Fixed